Building Theories from Case Study Research
This paper describes the process of inducting theory using case studies from specifying the research questions to reaching closure. Some process features, such as problem definition and construct validation, are similar to hypothesis-testing research. Others, such as within-case analysis and replication logic, are unique to the inductive, case-oriented process. Overall, the process described here is highly iterative and tightly linked to data. This research approach is especially appropriate in new topic areas. The resultant theory is often novel, testable, and empirically valid. Finally, the key criteria for evaluating this type of research are framebreaking insights, the tests of good theory (e.g., parsimony, logical coherence), and convincing grounding in the evidence.